pyproject.toml
[project]
name = "pldatacli"
version = "0.3.0"
description = "A data analysis cli tool using polars lazyframes"
readme = "README.md"
authors = [
    { name = "utkarshg1", email = "gaikwadujg@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "polars>=1.39.0",
    "pyyaml>=6.0.3",
    "rich>=14.3.3",
    "typer>=0.24.1",
]

[project.scripts]
pldatacli = "pldatacli.cli:app"

[dependency-groups]
dev = ["pytest>=8.0"]

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"
